+++ example/app.js
@@ -1,7 +1,6 @@
 const App = () => {
   const quill = React.useRef({});
   const [delta, setDelta] = React.useState('');
-  const [title, setTitle] = React.useState('1234');
   const initContent = '';
   const getQuill = (quillIns) => {
     quill.current = quillIns;
@@ -20,14 +19,14 @@ const App = () => {
         modules={{
           table: {},
           codeHighlight: true,
-          imageHandler: {
-            imgUploadApi: (formData) =>
-              // console.log(apiURL('uploadImg'))
-              request(apiURL('uploadImg'), ajaxFormPostOptions(formData)).then(
-                (response) => response.url,
-              ),
-            uploadFailCB: () => console.error('图片上传失败'),
-          },
+          // imageHandler: {
+          //   imgUploadApi: (formData) =>
+          //     // console.log(apiURL('uploadImg'))
+          //     request(apiURL('uploadImg'), ajaxFormPostOptions(formData)).then(
+          //       (response) => response.url,
+          //     ),
+          //   uploadFailCB: () => console.error('图片上传失败'),
+          // },
         }}
         getQuill={getQuill}
         content={
@@ -36,7 +35,6 @@ const App = () => {
         }
         onChange={quillChange}
         onFocus={(arg) => {}}
-        title={{ placeholder: 'Title', onChange: (title) => setTitle(title), value: title }}
         style={{ background: '#eee', height: 600 }}
       />
       <div style={{ height: 200 }}>{delta}</div>